refactor: remove 43 redundant (int) casts from PHPStan-typed arrays
## Why

With MYSQLI_OPT_INT_AND_FLOAT_NATIVE enabled, INT columns return PHP
int natively. When arrays are typed via @phpstan-type (StatsRow,
TeamWithStandingsRow, ProcessedTeamStats, etc.), PHPStan already knows
these fields are int. The (int) casts added visual noise without any
runtime or static-analysis benefit.

## Changes

- PlayerRegularSeasonTotalsView: remove 18 casts on StatsRow fields
- PlayerRegularSeasonAveragesView: remove 14 casts on StatsRow/CareerAveragesRow fields
- Team: remove 4 casts on TeamWithStandingsRow fields
- TeamOffDefStatsView: remove 4 casts on ProcessedTeamStats teamid
- NavigationView: remove 1 cast on inline-typed teamid
- GMContactListView: remove 1 cast on PHPDoc-typed teamid
- ContractListService: remove 1 cast on PlayerRow pid

## Test data fix

- FreeAgencyIntegrationTest: change HasMLE/HasLLE mock values from
  strings ('0'/'1') to integers (0/1) to match native types

+++ ibl5/classes/ContractList/ContractListService.php
@@ -50,7 +50,7 @@ class ContractListService implements ContractListServiceInterface
             $contractYears = $this->calculateContractYears($player);
 
             $contracts[] = [
-                'pid' => (int) ($player['pid'] ?? 0),
+                'pid' => $player['pid'] ?? 0,
                 'name' => $player['name'],
                 'pos' => $player['pos'],
                 'teamname' => $player['teamname'],
